ctranslate2::tokenizer

Trait Tokenizer

pub trait Tokenizer {
    // Required methods
    fn encode(&self, input: &str) -> Result<Vec<String>>;
    fn decode(&self, tokens: Vec<String>) -> Result<String>;
}

Required Methods

fn encode(&self, input: &str) -> Result<Vec<String>>

Encodes a given string into a sequence of tokens

fn decode(&self, tokens: Vec<String>) -> Result<String>

Decodes a given sequence of tokens back into a single string